LayoutTests/http/tests/media/media-source/video-media-source-play.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/media-resources/video-test.js"></script>
        <script src="/media/resources/media-source/webm/segment-info.js"></script>
        <script src="media-source.js"></script>
        <script>
            var segmentHelper = new MediaSourceTest.SegmentHelper(WebMSegmentInfo.testWebM);

            function onSourceOpen(event)
            {
                segmentHelper.addSourceBuffer();
                segmentHelper.appendInitSegment();

                // Append just enough segments to include at least 1 second of media data.
                var endIndex = segmentHelper.getMediaSegmentIndexForTimestamp(1);
                for (var i = 0; i <= endIndex; i++)
                    segmentHelper.appendMediaSegment(i);
                event.target.endOfStream();
            }

            function onLoad()
            {
                findMediaElement();

                mediaSource = new MediaSource();

                waitForEventAndFail('error');
                waitForEvent('webkitsourceopen', "", false, false, mediaSource);
                waitForEvent('playing');
                waitForEvent('webkitsourceended', "", false, false, mediaSource);
                waitForEventAndEnd('ended');

                mediaSource.addEventListener('webkitsourceopen', onSourceOpen);

                segmentHelper.init(video, function(success)
                {
                    if (!success) {
                        failTest("Failed to load segment data");
                        return;
                    }

                    MediaSourceTest.setSrcToMediaSourceTestURL(video);
                });
            }
        </script>
    </head>
    <body onload="onLoad()">
        <video autoplay> </video>
        <p>Tests normal playback case with MediaSource API</p>
    </body>
</html>
